update/delete/find_by_id repository and service updated

+++ model/service/bank_service.py
@@ -1,25 +1,40 @@
 from model import BankRepository
 
+
 class BankService:
     def __init__(self):
         self.repository = BankRepository()
 
     def save(self, bank):
-        self.repository.save(bank)
+        return self.repository.save(bank)
 
     def update(self, bank):
-        self.repository.update(bank)
+        bank = self.repository.find_by_id(bank.id)
+        if bank:
+            self.repository.update(bank)
+            return bank
+        else:
+            raise Exception("Bank Not Found !!!")
 
     def delete(self, bank_id):
-        self.repository.delete(bank_id)
+        bank = self.repository.find_by_id(bank_id)
+        if bank:
+            self.repository.delete(bank_id)
+            return bank
+        else:
+            raise Exception("Bank Not Found !!!")
 
     def find_all(self):
         return self.repository.find_all()
 
     def find_by_id(self, bank_id):
-        return self.repository.find_by_id(bank_id)
+        bank = self.repository.find_by_id(bank_id)
+        if bank:
+            return bank
+        else:
+            raise Exception("Bank Not Found !!!")
 
-    def find_by_name(self,name):
+    def find_by_name(self, name):
         return self.repository.find_by_name(name)
 
     def find_by_account(self, account):
