+++ model/service/product_service.py
@@ -6,19 +6,33 @@ class ProductService:
         self.repository = ProductRepository()
 
     def save(self, product):
-        self.repository.save(product)
+        return self.repository.save(product)
 
     def update(self, product):
-        self.repository.update(product)
+        product_result = self.repository.find_by_id(product.id)
+        if product_result:
+            self.repository.update(product)
+            return product
+        else:
+            raise Exception("Product Not Found !!!")
 
     def delete(self, product_id):
-        self.repository.delete(product_id)
+        product = self.repository.find_by_id(product_id)
+        if product:
+            self.repository.delete(product_id)
+            return product
+        else:
+            raise Exception("Product Not Found !!!")
 
     def find_all(self):
         return self.repository.find_all()
 
     def find_by_id(self, product_id):
-        return self.repository.find_by_id(product_id)
+        product = self.repository.find_by_id(product_id)
+        if product:
+            return product
+        else:
+            raise Exception("Product Not Found !!!")
 
     def find_by_name_and_brand(self,name,brand):
         return self.repository.find_by_name_and_brand(name,brand)
